@@ -1,9 +1,9 @@
-name: background_locator_2_example
+name: background_locator_cp_example
 description: Demonstrates how to use the background_locator_2 plugin.
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.8.0 <3.0.0"
+  sdk: ">=2.12.0 <3.2.0"
 
 dependencies:
   flutter:
@@ -17,11 +17,9 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  background_locator_2:
+  background_locator_cp:
     path: ../
 
-  path_provider: ^2.0.8
-  location_permissions: ^3.0.0+1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
