@@ -2,17 +2,17 @@ import 'dart:async';
 import 'dart:isolate';
 import 'dart:ui';
 
-import 'package:background_locator_2/background_locator.dart';
-import 'package:background_locator_2/location_dto.dart';
-import 'package:background_locator_2/settings/android_settings.dart';
-import 'package:background_locator_2/settings/ios_settings.dart';
-import 'package:background_locator_2/settings/locator_settings.dart';
+import 'package:background_locator_cp/background_locator.dart';
+import 'package:background_locator_cp/location_dto.dart';
+import 'package:background_locator_cp/settings/android_settings.dart';
+import 'package:background_locator_cp/settings/ios_settings.dart';
+import 'package:background_locator_cp/settings/locator_settings.dart';
 import 'package:flutter/material.dart';
-import 'package:location_permissions/location_permissions.dart';
 
-import 'file_manager.dart';
 import 'location_callback_handler.dart';
-import 'location_service_repository.dart';
+
+const String isolateName = 'LocatorIsolate';
+
 
 void main() => runApp(MyApp());
 
@@ -22,29 +22,30 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
+
   ReceivePort port = ReceivePort();
 
   String logStr = '';
-  bool isRunning;
-  LocationDto lastLocation;
+  late bool isRunning;
+  late LocationDto? lastLocation;
 
   @override
   void initState() {
     super.initState();
 
     if (IsolateNameServer.lookupPortByName(
-            LocationServiceRepository.isolateName) !=
+            isolateName) !=
         null) {
       IsolateNameServer.removePortNameMapping(
-          LocationServiceRepository.isolateName);
+          isolateName);
     }
 
     IsolateNameServer.registerPortWithName(
-        port.sendPort, LocationServiceRepository.isolateName);
+        port.sendPort, isolateName);
 
     port.listen(
       (dynamic data) async {
-        await updateUI(data);
+        print(data);
       },
     );
     initPlatformState();
@@ -55,88 +56,23 @@ class _MyAppState extends State<MyApp> {
     super.dispose();
   }
 
-  Future<void> updateUI(dynamic data) async {
-    final log = await FileManager.readLogFile();
-
-    LocationDto locationDto = (data != null) ? LocationDto.fromJson(data) : null;
-    await _updateNotificationText(locationDto);
-
-    setState(() {
-      if (data != null) {
-        lastLocation = locationDto;
-      }
-      logStr = log;
-    });
-  }
-
-  Future<void> _updateNotificationText(LocationDto data) async {
-    if (data == null) {
-      return;
-    }
-
-    await BackgroundLocator.updateNotificationText(
-        title: "new location received",
-        msg: "${DateTime.now()}",
-        bigMsg: "${data.latitude}, ${data.longitude}");
-  }
 
   Future<void> initPlatformState() async {
     print('Initializing...');
     await BackgroundLocator.initialize();
-    logStr = await FileManager.readLogFile();
     print('Initialization done');
     final _isRunning = await BackgroundLocator.isServiceRunning();
     setState(() {
       isRunning = _isRunning;
     });
+
+    _startLocator();
     print('Running ${isRunning.toString()}');
   }
 
   @override
   Widget build(BuildContext context) {
-    final start = SizedBox(
-      width: double.maxFinite,
-      child: ElevatedButton(
-        child: Text('Start'),
-        onPressed: () {
-          _onStart();
-        },
-      ),
-    );
-    final stop = SizedBox(
-      width: double.maxFinite,
-      child: ElevatedButton(
-        child: Text('Stop'),
-        onPressed: () {
-          onStop();
-        },
-      ),
-    );
-    final clear = SizedBox(
-      width: double.maxFinite,
-      child: ElevatedButton(
-        child: Text('Clear Log'),
-        onPressed: () {
-          FileManager.clearLogFile();
-          setState(() {
-            logStr = '';
-          });
-        },
-      ),
-    );
-    String msgStatus = "-";
-    if (isRunning != null) {
-      if (isRunning) {
-        msgStatus = 'Is running';
-      } else {
-        msgStatus = 'Is not running';
-      }
-    }
-    final status = Text("Status: $msgStatus");
 
-    final log = Text(
-      logStr,
-    );
 
     return MaterialApp(
       home: Scaffold(
@@ -146,12 +82,6 @@ class _MyAppState extends State<MyApp> {
         body: Container(
           width: double.maxFinite,
           padding: const EdgeInsets.all(22),
-          child: SingleChildScrollView(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: <Widget>[start, stop, clear, status, log],
-            ),
-          ),
         ),
       ),
     );
@@ -163,44 +93,6 @@ class _MyAppState extends State<MyApp> {
     setState(() {
       isRunning = _isRunning;
     });
-  }
-
-  void _onStart() async {
-    if (await _checkLocationPermission()) {
-      await _startLocator();
-      final _isRunning = await BackgroundLocator.isServiceRunning();
-
-      setState(() {
-        isRunning = _isRunning;
-        lastLocation = null;
-      });
-    } else {
-      // show error
-    }
-  }
-
-  Future<bool> _checkLocationPermission() async {
-    final access = await LocationPermissions().checkPermissionStatus();
-    switch (access) {
-      case PermissionStatus.unknown:
-      case PermissionStatus.denied:
-      case PermissionStatus.restricted:
-        final permission = await LocationPermissions().requestPermissions(
-          permissionLevel: LocationPermissionLevel.locationAlways,
-        );
-        if (permission == PermissionStatus.granted) {
-          return true;
-        } else {
-          return false;
-        }
-        break;
-      case PermissionStatus.granted:
-        return true;
-        break;
-      default:
-        return false;
-        break;
-    }
   }
 
   Future<void> _startLocator() async{
