@@ -1,30 +1,30 @@
 import 'dart:async';
+import 'dart:isolate';
+import 'dart:ui';
 
-import 'package:background_locator_2/location_dto.dart';
+import 'package:background_locator_cp/location_dto.dart';
 
-import 'location_service_repository.dart';
+import 'main.dart';
+
+
 
 @pragma('vm:entry-point')
 class LocationCallbackHandler {
+
   @pragma('vm:entry-point')
   static Future<void> initCallback(Map<dynamic, dynamic> params) async {
-    LocationServiceRepository myLocationCallbackRepository =
-        LocationServiceRepository();
-    await myLocationCallbackRepository.init(params);
+
   }
 
   @pragma('vm:entry-point')
   static Future<void> disposeCallback() async {
-    LocationServiceRepository myLocationCallbackRepository =
-        LocationServiceRepository();
-    await myLocationCallbackRepository.dispose();
+
   }
 
   @pragma('vm:entry-point')
   static Future<void> callback(LocationDto locationDto) async {
-    LocationServiceRepository myLocationCallbackRepository =
-        LocationServiceRepository();
-    await myLocationCallbackRepository.callback(locationDto);
+    SendPort? send = IsolateNameServer.lookupPortByName(isolateName);
+    send?.send("Aiai amor");
   }
 
   @pragma('vm:entry-point')
