@@ -1,10 +1,10 @@
 import 'dart:io';
 import 'dart:ui';
 
-import 'package:background_locator_2/keys.dart';
-import 'package:background_locator_2/location_dto.dart';
-import 'package:background_locator_2/settings/android_settings.dart';
-import 'package:background_locator_2/settings/ios_settings.dart';
+import 'package:background_locator_cp/keys.dart';
+import 'package:background_locator_cp/location_dto.dart';
+import 'package:background_locator_cp/settings/android_settings.dart';
+import 'package:background_locator_cp/settings/ios_settings.dart';
 
 class SettingsUtil {
   static Map<String, dynamic> getArgumentsMap(
